WIP: Support NameIDFormats:1/0 attribute
Manage can send the NameIDFormats (plural!) attribute in the
metadatafields. The dashboard should not drop that property but send
it back when publising an entity.

This commit is probaby not very useful, the dashboard should handle
metadata fields in a generic way so we don't have to support each
attribute explicitly, it should just remember all attributes and send
them back.

See: https://www.pivotaltracker.com/story/show/153810318